test/fixtures/wpt/resource-timing/input-sequence-of-events.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test the sequence of events when reporting input timing.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
  async_test(t => {
    const input = document.createElement('input');
    input.type = "image";
    const absoluteURL = new URL('resources/blue.png', location.href).toString();
    t.add_cleanup(() => input.remove());
    input.addEventListener('load', t.step_func(() => {
      assert_equals(performance.getEntriesByName(absoluteURL).length, 1);
      t.done();
    }));
    input.src = absoluteURL;
    document.body.appendChild(input);
  }, "An image input element should receive its load event after the ResourceTiming entry is available");

</script>
